# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cabo Fit Pass API
  version: 1.0.0
  description: |
    **Cabo Fit Pass** is a production-ready fitness booking platform built for the Los Cabos market. 
    
    This API provides a comprehensive suite of endpoints for:
    - **User Authentication & Profiles** - Secure user management with NextAuth.js
    - **Class Booking System** - Credit-based booking with atomic transactions
    - **Payment Processing** - Stripe integration with webhook support
    - **Admin Management** - Gym and studio administration
    - **Health Monitoring** - System health and performance tracking
    
    ## Authentication
    The API uses JWT-based authentication via NextAuth.js. Include the session token in requests to protected endpoints.
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse:
    - **Standard endpoints**: 100 requests per minute per IP
    - **Payment endpoints**: 20 requests per minute per user
    - **Admin endpoints**: 50 requests per minute per admin user
    
    ## Error Handling
    All endpoints return consistent error responses with:
    - HTTP status codes following REST conventions
    - Structured error objects with `success: false`
    - Detailed error messages for debugging (development only)
    - Request IDs for tracing (via headers)

  contact:
    name: Cabo Fit Pass API Support
    url: https://github.com/your-repo/cabo-fit-pass
    email: support@cabofitpass.com
  license:
    name: MIT
...
servers:
...
security:
  - BearerAuth: []
  - SessionAuth: []
paths:
...
  /credits:
    get:
      tags:
        - Credits
      summary: Get user's credit balance
      description: |
        Returns the current credit balance for the authenticated user.
        
        **Caching**: Results cached for 5 minutes for performance
      operationId: getUserCredits
      responses:
        '200':
          description: Credit balance retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          credits:
                            type: integer
                            description: Current credit balance
              example:
                success: true
                data:
                  credits: 25
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User profile not found

    post:
      tags:
        - Credits
      summary: Add credits to user account
      description: |
        Adds credits to a user's account after payment verification.
        
        **Security**: 
        - Verifies payment with Stripe before adding credits
        - User can only add credits to their own account
        - All transactions are logged for audit
        
        **Rate Limiting**: 10 credit purchases per hour per user
      operationId: addUserCredits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_email
                - credits_to_add
              properties:
                user_email:
                  type: string
                  format: email
                  description: Email of user receiving credits
                credits_to_add:
                  type: integer
                  minimum: 1
                  maximum: 100
                  description: Number of credits to add
                payment_intent_id:
                  type: string
                  description: Stripe payment intent ID for verification
            example:
              user_email: "john@example.com"
              credits_to_add: 10
              payment_intent_id: "pi_1234567890"
      responses:
        '200':
          description: Credits added successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          success:
                            type: boolean
                          new_credits:
                            type: integer
                            description: New total credit balance
                          added_credits:
                            type: integer
                            description: Number of credits added
        '400':
          description: Invalid credit amount or payment verification failed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Cannot add credits for different user
...
  /admin/gyms:
    get:
      tags:
        - Admin
      summary: List all gyms
      description: |
        Retrieves all gyms in the system with their details and connection status.
        
        **Admin Only**: Requires admin role
      operationId: listGyms
      responses:
        '200':
          description: Gyms retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Gym'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Admin
      summary: Create new gym
      description: |
        Creates a new gym in the system.
        
        **Admin Only**: Requires admin role
      operationId: createGym
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                  description: Gym name
                email:
                  type: string
                  format: email
                  description: Gym contact email
                phone:
                  type: string
                  description: Gym phone number
                location:
                  type: string
                  description: Gym location/address
      responses:
        '200':
          description: Gym created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Gym'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication
    
    SessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
  schemas:
...
    Gym:
      type: object
      required:
        - id
        - name
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique gym identifier
        name:
          type: string
          maxLength: 100
          description: Gym name
        location:
          type: string
          nullable: true
          description: Gym location
        email:
          type: string
          format: email
          nullable: true
          description: Gym contact email
        phone:
          type: string
          nullable: true
          description: Gym phone number
        stripe_connect_id:
          type: string
          nullable: true
          description: Stripe Connect account ID
        connect_onboarded_at:
          type: string
          format: date-time
          nullable: true
          description: Stripe Connect onboarding completion timestamp
        created_at:
          type: string
          format: date-time
          description: Gym creation timestamp
        updated_at:
          type: string
          format: date-time
          nullable: true
          description: Last update timestamp
      example:
        id: "gym-123"
        name: "Los Cabos Fitness Center"
        location: "Downtown Cabo San Lucas"
        email: "info@cabofitness.com"
        phone: "+52-624-123-4567"
        created_at: "2024-01-15T08:00:00.000Z"

    # Response Schemas
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          type: object
          nullable: true
          description: Response data (varies by endpoint)
        error:
          type: string
          nullable: true
          description: Error message if success is false
        message:
          type: string
          nullable: true
          description: Additional information or success message
      example:
        success: true
        message: "Operation completed successfully"
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
          description: Error message
        code:
          type: string
          nullable: true
          description: Error code for programmatic handling
        details:
          type: object
          nullable: true
          description: Additional error details
      example:
        success: false
        error: "Validation failed"
        code: "VALIDATION_ERROR"
        details:
          field: "email"
          message: "Invalid email format"
  responses:
    BadRequest:
      description: Bad Request - Invalid input or request format
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Invalid request parameters"
            code: "BAD_REQUEST"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Authentication required"
            code: "UNAUTHORIZED"
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Insufficient permissions"
            code: "FORBIDDEN"
...